Updated README.md and implemented API stub
