+++ package.json
@@ -2,7 +2,7 @@
   "name": "limbus-buildgen",
   "version": "0.1.1",
   "description": "A \"build anywhere\" build system.",
-  "main": "limbus-buildgen",
+  "main": "source/limbus-buildgen.js",
   "directories": {},
   "bin": {
     "limbus-buildgen": "./limbus-buildgen"
@@ -26,7 +26,7 @@
     "coverage": "mocha --require blanket -R html-cov \"**/*.unit.js\" \"**/*.feature.js\" > coverage.html",
     "unit-test": "mocha \"**/*.unit.js\" ||echo",
     "integration-test": "npm run-script build && mocha -t 20000 \"**/*.feature.js\" ||echo",
-    "pack-js": "webpack source/limbus-buildgen.js packed.js --config webpack.config.js && node -e \"var fs = require('fs'); console.log('module = {exports: {}};' + fs.readFileSync('packed.js').toString());\" | uglifyjs - --compress --mangle | stream-to-c-header-char-array javascript_source > source/packed_javascript_source.h",
+    "pack-js": "webpack source/limbus-buildgen-cli.js packed.js --config webpack.config.js && node -e \"var fs = require('fs'); console.log('module = {exports: {}};' + fs.readFileSync('packed.js').toString());\" | uglifyjs - --compress --mangle | stream-to-c-header-char-array javascript_source > source/packed_javascript_source.h",
     "build": "npm run-script pack-js && make",
     "install": "make"
   },
